refactor(discord): split queued cancellation seams
Move the primary-ID adapter and queue cancellation helpers into their owning
modules so the safe cancellation control stays within giant-file limits.
