+++ ARCHITECTURE.md
@@ -1150,6 +1150,7 @@ src/
 │   │   ├── dispatch_reservation.rs
 │   │   ├── overflow.rs
 │   │   ├── pending_queue_persistence.rs
+│   │   ├── queue_cancellation.rs
 │   │   ├── registry_purge.rs
 │   │   ├── source_generation.rs
 │   │   └── turn_finished_signal.rs
