+++ src/services/discord/queue_io.rs
@@ -85,6 +85,23 @@ async fn idle_queue_kick_hook_outcome_for_tests(
     hook?(shared, provider, channel_id, reason).await
 }
 
+pub(in crate::services::discord) async fn mailbox_cancel_queued_primary_message(
+    shared: &SharedData,
+    provider: &ProviderKind,
+    channel_id: ChannelId,
+    message_id: MessageId,
+) -> Option<Intervention> {
+    let result: CancelQueuedMessageResult = shared
+        .mailbox(channel_id)
+        .cancel_queued_primary_message(
+            message_id,
+            queue_persistence_context(shared, provider, channel_id),
+        )
+        .await;
+    apply_queue_exit_feedback(shared, channel_id, &result.queue_exit_events).await;
+    result.removed
+}
+
 pub(super) async fn with_post_enqueue_idle_queue_kick_suppressed<F>(future: F) -> F::Output
 where
     F: std::future::Future,
